DeleteProductHandlerda ürün silinme kontrolü yapıldı

+++ Server/Server.Application/Features/Products/DeleteProductById/DeleteProductByIdCommandHandler.cs
@@ -7,6 +7,9 @@ namespace Server.Application.Features.Products.DeleteProductById
 {
     public sealed class DeleteProductByIdCommandHandler(
         IProductRepository productRepository,
+        IInvoiceDetailRepository invoiceDetailRepository,
+        IRecipeDetailRepository recipeDetailRepository,
+        IOrderDetailRepository orderDetailRepository,
         IUnitOfWork unitOfWork) : IRequestHandler<DeleteProductByIdCommand, Result<string>>
     {
         public async Task<Result<string>> Handle(DeleteProductByIdCommand request, CancellationToken cancellationToken)
@@ -18,6 +21,22 @@ namespace Server.Application.Features.Products.DeleteProductById
                 return Result<string>.Failure("Ürün Bulunamadı");
             }
 
+            bool invoiceControl = await invoiceDetailRepository.AnyAsync(p => p.ProductId == request.Id);
+            if (invoiceControl)
+            {
+                return Result<string>.Failure("Ürün Faturalarda Kullanılmaktadır.Silinemez");
+            }
+
+            bool recipeControl = await recipeDetailRepository.AnyAsync(p => p.ProductId == request.Id);
+            if (invoiceControl)
+            {
+                return Result<string>.Failure("Ürün Reçetelrde Kullanılmaktadır.Silinemez");
+            }
+            bool orderControl = await orderDetailRepository.AnyAsync(p => p.ProductId == request.Id);
+            if (orderControl)
+            {
+                return Result<string>.Failure("Ürün Siparişlerde Kullanılmaktadır.Silinemez");
+            }
             productRepository.Delete(product);
             await unitOfWork.SaveChangesAsync();
